+++ src/dakara_feeder/song.py
@@ -1,6 +1,6 @@
 import logging
 
-from dakara_feeder.metadata_parser import FFProbeMetadataParser
+from dakara_feeder.metadata_parser import FFProbeMetadataParser, MediaParseError
 from dakara_feeder.subtitle_parser import Pysubs2SubtitleParser, SubtitleParseError
 
 
@@ -113,8 +113,14 @@ class BaseSong:
         Returns:
             float: Duration of the song in seconds.
         """
-        parser = FFProbeMetadataParser.parse(self.base_directory / self.video_path)
-        return parser.get_duration().total_seconds()
+        try:
+            parser = FFProbeMetadataParser.parse(self.base_directory / self.video_path)
+            return parser.get_duration().total_seconds()
+
+        except MediaParseError as error:
+            logger.error("Duration not parsed: {}".format(error))
+            # Set duration to zero when duration can't be extracted
+            return 0
 
     def get_artists(self):
         """Get the list of artists