+++ tests/test_song.py
@@ -5,7 +5,7 @@ from unittest.mock import patch
 from path import Path
 
 from dakara_feeder.directory_lister import SongPaths
-from dakara_feeder.metadata_parser import FFProbeMetadataParser
+from dakara_feeder.metadata_parser import FFProbeMetadataParser, MediaParseError
 from dakara_feeder.song import BaseSong
 from dakara_feeder.subtitle_parser import Pysubs2SubtitleParser, SubtitleParseError
 
@@ -42,4 +42,32 @@ class BaseSongTestCase(TestCase):
 
         self.assertListEqual(
             logger.output, ["ERROR:dakara_feeder.song:Lyrics not parsed: invalid"]
+        )
+
+    @patch.object(Pysubs2SubtitleParser, "parse", autoset=True)
+    @patch.object(FFProbeMetadataParser, "parse", autoset=True)
+    def test_metadata_parser_error(self, mocked_metadata_parse, mocked_subtitle_parse):
+        """Test an invalid video file raises no exception but logs error
+        """
+        # setup mocks
+        mocked_metadata_parse.side_effect = MediaParseError("invalid")
+        mocked_subtitle_parse.return_value.get_lyrics.return_value = ""
+
+        # create paths
+        paths = SongPaths(Path("file.mp4"), Path("file.ass"))
+
+        # create BaseSong instance
+        song = BaseSong(Path("/base-dir"), paths)
+
+        # get song representation
+        with self.assertLogs("dakara_feeder.song") as logger:
+            representation = song.get_representation()
+
+        # check duration defaults to zero
+        self.assertEqual(representation["duration"], 0)
+
+        # assert logs
+
+        self.assertListEqual(
+            logger.output, ["ERROR:dakara_feeder.song:Duration not parsed: invalid"]
         )