+++ CHANGELOG.md
@@ -32,7 +32,8 @@
 
 ### Fixed
 
-- An invalid subtitle file no longer halts the feed process but only logs a warning.
+- A subtitle file from which lyrics can't be parsed no longer halts the feed process but only logs an error.
+- A video file from which duration can't be extracted no longer halts the feed process but only logs an error.
 
 ## 1.5.0 - 2019-12-05